Move custom_fields to AbstractRequest, update to java 8
