+++ .travis.yml
@@ -1,1 +1,4 @@
 language: java
+
+jdk:
+  - openjdk8
