+++ pom.xml
@@ -5,7 +5,7 @@
     <groupId>com.hellosign</groupId>
     <artifactId>hellosign-java-sdk</artifactId>
     <packaging>jar</packaging>
-    <version>4.0.12</version>
+    <version>4.1.0</version>
     <name>HelloSign Java SDK</name>
     <url>https://github.com/hellosign/hellosign-java-sdk</url>
     <properties>
@@ -169,6 +169,14 @@
                     </descriptorRefs>
                 </configuration>
             </plugin>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-compiler-plugin</artifactId>
+            <configuration>
+              <source>8</source>
+              <target>8</target>
+            </configuration>
+          </plugin>
         </plugins>
     </build>
     <reporting>
